Home Automation E-Mail notifications
This is a full home automation project implemented on this architecture
house using the Arduino Uno,ESP8266 and different sensor with E-Mail
notifications using python. You can use the 3g-GSM/GPRS module for SMS
notifications but this is the cheapest solution.
The logic and the
process by which this project has been built doesn't have much
difference with respect to the real world or the industry.
All the codes
are there and have been shared step by step.